% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_factors.R
\name{get_factors}
\alias{get_factors}
\title{Returns a stars object with projection factors calculated at each pixel as variables}
\usage{
get_factors(s)
}
\arguments{
\item{s}{stars object}
}
\value{
stars object
}
\description{
Returns a stars object with projection factors calculated at each pixel as variables
}
