% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_factors.R
\name{interpolate_xy}
\alias{interpolate_xy}
\title{Extract value using a bilinear interpolation}
\usage{
interpolate_xy(s, x, y)
}
\arguments{
\item{s}{stars object}

\item{x}{x coordinate to be interpolated}

\item{y}{y coordinate to be interpolated}
}
\value{
interpolated value
}
\description{
Extract value using a bilinear interpolation
}
