% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_factors.R
\name{st_convolve}
\alias{st_convolve}
\title{Calculate statistics using floating window}
\usage{
st_convolve(s, stats = "mean", size = 3, adaptive = FALSE)
}
\arguments{
\item{s}{a raster stars object}

\item{stats}{a vector of stats names}

\item{size}{a size of the kernel. 3 by defaulr}

\item{adaptive}{should a kernel be adaptive to map projection distorions?}
}
\value{
a stars object
}
\description{
Calculate statistics using floating window
}
