% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_factors.R
\name{st_convolve}
\alias{st_convolve}
\title{Filter stars object with selected kernel}
\usage{
st_convolve(s, kernel = "mean", size = c(3, 3), fun = NULL)
}
\arguments{
\item{s}{a raster stars object}

\item{kernel}{string (name of the filter), matrix (with values)}

\item{size}{}
}
\value{
matrix
}
\description{
Filter stars object with selected kernel
}
